pyproject.toml
[tool.poetry]
name = "cms-parser"
version = "1.0.0"
description = "A package for parsing CMS data"
authors = ["Naomi Lee"]
license = "MIT License"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.6"
black = "23.3.0"
camelot-py = "0.11.0"
certifi = "2022.12.7"
cffi = "1.15.1"
chardet = "5.1.0"
charset-normalizer = "3.1.0"
click = "8.1.3"
contourpy = "1.0.7"
cryptography = "40.0.2"
cycler = "0.11.0"
distro = "1.8.0"
et-xmlfile = "1.1.0"
fonttools = "4.39.3"
ghostscript = "0.7"
idna = "3.4"
iniconfig = "2.0.0"
kiwisolver = "1.4.4"
matplotlib = "3.7.1"
mypy-extensions = "1.0.0"
numpy = "1.24.3"
opencv-python = "4.7.0.72"
openpyxl = "3.1.2"
packaging = "23.1"
pandas = "2.0.1"
pathspec = "0.11.1"
pdfminer-six = "20221105"
pillow = "9.5.0"
pip = "23.1.2"
platformdirs = "3.5.0"
pluggy = "1.0.0"
pycparser = "2.21"
pyparsing = "3.0.9"
pytest = "7.3.1"
python-dateutil = "2.8.2"
pytz = "2023.3"
setuptools = "67.7.2"
six = "1.16.0"
soupsieve = "2.4.1"
tabulate = "0.9.0"
tzdata = "2023.3"
urllib3 = "1.26.15"
vermin = "1.5.1"
wheel = "0.40.0"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[options.entry_points]
console_scripts = {cms_parser = "src.cms_parser.main:main"}
